make settlement works

+++ protocol-units/settlement/mcr/config/src/lib.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 const MCR_CONTRACT_ADDRESS: &str = "0xBf7c7AE15E23B2E19C7a1e3c36e245A71500e181";
 const DEFAULT_BATCH_TIMEOUT_MILLIS: u64 = 2000;
 const DEFAULT_TX_SEND_RETRIES: u32 = 10;
-const DEFAULT_GAS_LIMIT: u128 = 10_000_000_000_000_000;
+const DEFAULT_GAS_LIMIT: u64 = 10_000_000_000_000_000;
 
 /// Configuration of the MCR settlement client.
 ///
@@ -23,7 +23,7 @@ pub struct Config {
 	#[serde(default = "default_mcr_contract_address")]
 	pub mcr_contract_address: String,
 	#[serde(default = "default_gas_limit")]
-	pub gas_limit: u128,
+	pub gas_limit: u64,
 	/// Timeout for batching blocks, in milliseconds
 	#[serde(default = "default_batch_timeout")]
 	pub batch_timeout: u64,
@@ -40,7 +40,7 @@ fn default_mcr_contract_address() -> String {
 	MCR_CONTRACT_ADDRESS.into()
 }
 
-const fn default_gas_limit() -> u128 {
+const fn default_gas_limit() -> u64 {
 	DEFAULT_GAS_LIMIT
 }
